refactored based on pr feedback
